Create states for the error simulator
